Adjusted the file structure so it more closely matches what we have been taught. - Static and Templates Folders now stored in app folder.
