+++ README.md
@@ -1,2 +1,4 @@
 # AWD_Group_Project
 Repository for the UWA Agile Web Development 2026 Group Project
+
+pip install -r requirements.txt
