+++ app/static/Create_Recipe/editAppliances.js
