+++ app/static/Create_Recipe/getRecipeInfo.js
