+++ app/static/Create_Recipe/populateRecipeDetails.js
