+++ app/static/Explore/displayRecipes.js
