+++ app/static/Explore/displaySpotlight.js
